Recompiled, x64 build, general cleanup, and more
Recompiled with the newest version of Lazarus (Lazarus 1.6.4) with
better Unicode support. Compiled a x64 executable for those who prefer
it. General cleanup of files. Included missing files that were not
correctly in sync during the initial upload (this include the necessary
XYplorer script to use XYplorer in conjunction with XYZback). Updated
Zback to latest version 2.85.0.c from December 2016. Very minor changes
in code. Version bump to final 1.0 as it's stable. This may be the last
update since the developer of XYplorer is planning on adding syncing and
mirroring capabilities to XYplorer in the near future. Read 'Readme.md'
for further details. Greatly updated 'Readme.md' for clarity, other
notes, and how to build 'XYZback'.